// types/casehistory/casehistoryform.ts
export interface CaseHistoryFormData {
    user_id: string;
    session_id: string;
    session_type: "video" | "chat";
    counsellor_id: string;
    address: string;
    education: string;
    occupation: string;
    marital_status: string;
    family_type: string;
    family_members: string;
    identification_mark_1: string;
    identification_mark_2: string;
    reliability: string;
    previous_consultation: boolean;
    previous_consultation_details: string;
    chief_complaints: string;
    present_illness_history: string;
    treatment_history: string;
    past_illness_history: string;
    family_history: string;
    childhood_disorders: string;
    home_atmosphere: string;
    scholastic_activities: string;
    sexual_marital_history: string;
    premorbid_personality: string;
    general_appearance: string;
    attitude: string;
    motor_behavior: string;
    speech: string;
    attention_concentration: string;
    orientation: string;
    memory: string;
    abstract_ability: string;
    general_information: string;
    calculation: string;
    intelligence: string;
    mood_affect: string;
    thought_stream: string;
    thought_form: string;
    thought_possession: string;
    thought_content: string;
    advice_activities: string;
    follow_up_session: string;
  }